app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-multipart==0.0.6
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pandas==2.1.3
openpyxl==3.1.2
